+++ core/test/config/filters-test.js
@@ -6,11 +6,11 @@
 
 import log from 'lighthouse-logger';
 
-import {Audit as BaseAudit} from '../../../audits/audit.js';
-import BaseGatherer from '../../../gather/base-gatherer.js';
-import {defaultSettings, defaultNavigationConfig} from '../../../config/constants.js';
-import * as filters from '../../../config/filters.js';
-import {initializeConfig} from '../../../config/config.js';
+import {Audit as BaseAudit} from '../../audits/audit.js';
+import BaseGatherer from '../../gather/base-gatherer.js';
+import {defaultSettings, defaultNavigationConfig} from '../../config/constants.js';
+import * as filters from '../../config/filters.js';
+import {initializeConfig} from '../../config/config.js';
 
 describe('Fraggle Rock Config Filtering', () => {
   const snapshotGatherer = new BaseGatherer();
