+++ core/test/config/validation-test.js
@@ -4,12 +4,12 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {defaultSettings, defaultNavigationConfig} from '../../../config/constants.js';
-import defaultConfig from '../../../config/default-config.js';
-import {Audit as BaseAudit} from '../../../audits/audit.js';
-import BaseFRGatherer from '../../../gather/base-gatherer.js';
-import {Gatherer as BaseLegacyGatherer} from '../../../gather/gatherers/gatherer.js';
-import * as validation from '../../../config/validation.js';
+import {defaultSettings, defaultNavigationConfig} from '../../config/constants.js';
+import defaultConfig from '../../config/default-config.js';
+import {Audit as BaseAudit} from '../../audits/audit.js';
+import BaseFRGatherer from '../../gather/base-gatherer.js';
+import {Gatherer as BaseLegacyGatherer} from '../../gather/gatherers/gatherer.js';
+import * as validation from '../../config/validation.js';
 
 /** @typedef {LH.Gatherer.GathererMeta['supportedModes']} SupportedModes */
