+++ core/test/gather/base-artifacts-test.js
@@ -4,10 +4,10 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {getBaseArtifacts, finalizeArtifacts} from '../../../gather/base-artifacts.js';
-import {initializeConfig} from '../../../config/config.js';
+import {getBaseArtifacts, finalizeArtifacts} from '../../gather/base-artifacts.js';
+import {initializeConfig} from '../../config/config.js';
 import {createMockDriver} from './mock-driver.js';
-import {LighthouseError} from '../../../lib/lh-error.js';
+import {LighthouseError} from '../../lib/lh-error.js';
 
 function getMockDriverForArtifacts() {
   const driverMock = createMockDriver();
