+++ core/test/gather/base-gatherer-test.js
@@ -4,8 +4,8 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import Gatherer from '../../../gather/base-gatherer.js';
-import {fnAny} from '../../test-utils.js';
+import Gatherer from '../../gather/base-gatherer.js';
+import {fnAny} from '../test-utils.js';
 
 /** @type {any} */
 const fakeParam = {};
