+++ core/test/gather/driver/network-monitor-test.js
@@ -4,7 +4,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {createMockCdpSession} from '../../fraggle-rock/gather/mock-driver.js';
+import {createMockCdpSession} from '../mock-driver.js';
 import {NetworkMonitor} from '../../../gather/driver/network-monitor.js';
 import {NetworkRequest} from '../../../lib/network-request.js';
 import {networkRecordsToDevtoolsLog} from '../../network-records-to-devtools-log.js';
