+++ core/test/gather/driver/service-workers-test.js
@@ -4,7 +4,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {createMockSession} from '../../fraggle-rock/gather/mock-driver.js';
+import {createMockSession} from '../mock-driver.js';
 import {makePromiseInspectable, flushAllTimersAndMicrotasks, timers} from '../../test-utils.js';
 import * as serviceWorkers from '../../../gather/driver/service-workers.js';
