+++ core/test/gather/driver/storage-test.js
@@ -4,7 +4,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {createMockSession} from '../../fraggle-rock/gather/mock-driver.js';
+import {createMockSession} from '../mock-driver.js';
 import * as storage from '../../../gather/driver/storage.js';
 
 let sessionMock = createMockSession();
