+++ core/test/gather/gatherers/dobetterweb/optimized-images-test.js
@@ -5,7 +5,7 @@
  */
 
 import OptimizedImages from '../../../../gather/gatherers/dobetterweb/optimized-images.js';
-import {createMockContext} from '../../../fraggle-rock/gather/mock-driver.js';
+import {createMockContext} from '../../../gather/mock-driver.js';
 
 let context = createMockContext();
 let optimizedImages;
