+++ core/test/gather/gatherers/dobetterweb/response-compression-test.js
@@ -4,7 +4,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {createMockContext, mockDriverSubmodules} from '../../../fraggle-rock/gather/mock-driver.js';
+import {createMockContext, mockDriverSubmodules} from '../../../gather/mock-driver.js';
 // import ResponseCompression from '../../../../gather/gatherers/dobetterweb/response-compression.js';
 
 // Some imports needs to be done dynamically, so that their dependencies will be mocked.
