+++ core/test/gather/gatherers/inspector-issues-test.js
@@ -6,7 +6,7 @@
 
 import InspectorIssues from '../../../gather/gatherers/inspector-issues.js';
 import {NetworkRequest} from '../../../lib/network-request.js';
-import {createMockContext} from '../../fraggle-rock/gather/mock-driver.js';
+import {createMockContext} from '../mock-driver.js';
 import {flushAllTimersAndMicrotasks, timers} from '../../test-utils.js';
 import {networkRecordsToDevtoolsLog} from '../../network-records-to-devtools-log.js';
