+++ core/test/gather/gatherers/main-document-content-test.js
@@ -6,7 +6,7 @@
 
 import MainDocumentContent from '../../../gather/gatherers/main-document-content.js';
 import {NetworkRecorder} from '../../../lib/network-recorder.js';
-import {createMockContext} from '../../fraggle-rock/gather/mock-driver.js';
+import {createMockContext} from '../mock-driver.js';
 import {getURLArtifactFromDevtoolsLog, readJson} from '../../test-utils.js';
 
 const devtoolsLog = readJson('../../fixtures/traces/lcp-m78.devtools.log.json', import.meta);
