+++ core/test/scenarios/cross-origin-test-pptr.js
@@ -4,9 +4,9 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import * as lighthouse from '../../../api.js';
+import * as lighthouse from '../../api.js';
 import {createTestState} from './pptr-test-utils.js';
-import {LH_ROOT} from '../../../../root.js';
+import {LH_ROOT} from '../../../root.js';
 
 /* eslint-env browser */
