+++ core/test/user-flow-test.js
@@ -7,28 +7,28 @@
 import jestMock from 'jest-mock';
 import * as td from 'testdouble';
 
-import {Runner} from '../../runner.js';
+import {Runner} from '../runner.js';
 import {createMockPage, mockRunnerModule} from './gather/mock-driver.js';
 // import UserFlow from '../../user-flow.js';
 
 // Some imports needs to be done dynamically, so that their dependencies will be mocked.
 // See: https://jestjs.io/docs/ecmascript-modules#differences-between-esm-and-commonjs
 //      https://github.com/facebook/jest/issues/10025
-/** @type {typeof import('../../user-flow.js').UserFlow} */
+/** @type {typeof import('../user-flow.js').UserFlow} */
 let UserFlow;
-/** @type {typeof import('../../user-flow.js')['auditGatherSteps']} */
+/** @type {typeof import('../user-flow.js')['auditGatherSteps']} */
 let auditGatherSteps;
 
 before(async () => {
-  ({UserFlow, auditGatherSteps} = await import('../../user-flow.js'));
+  ({UserFlow, auditGatherSteps} = await import('../user-flow.js'));
 });
 
 const snapshotModule = {snapshotGather: jestMock.fn()};
-await td.replaceEsm('../../gather/snapshot-runner.js', snapshotModule);
+await td.replaceEsm('../gather/snapshot-runner.js', snapshotModule);
 const navigationModule = {navigationGather: jestMock.fn()};
-await td.replaceEsm('../../gather/navigation-runner.js', navigationModule);
+await td.replaceEsm('../gather/navigation-runner.js', navigationModule);
 const timespanModule = {startTimespanGather: jestMock.fn()};
-await td.replaceEsm('../../gather/timespan-runner.js', timespanModule);
+await td.replaceEsm('../gather/timespan-runner.js', timespanModule);
 
 const mockRunner = await mockRunnerModule();
